Changing type of file_reference column to bigint to fix error due to size of input numbers.

+++ src/asset_folder_importer/asset_folder_importer_database.sql
@@ -153,7 +153,7 @@ CREATE TABLE prelude_clips (
     import_date timestamp without time zone,
     project uuid,
     start_time bigint,
-    file_reference integer,
+    file_reference bigint,
     parent_id integer
 );
